# api_gateway/app/dto/house_keeper.py
from pydantic import BaseModel
import uuid
class HouseKeeper(BaseModel):
    id: uuid.UUID
    firstName: str
    lastName: str
    phone: str

class HouseKeeperCreate(BaseModel):
    firstName: str
    lastName: str
    phone: str